requirements.txt
torch
gymnasium
flappy-bird-gymnasium
PyYAML
pygame